## Description
Opening of the form view for invoices is slow on large databases. It's due to the presence of the `l10n_in_withhold_move_ids` in the form view. It's inverse `l10n_in_withholding_ref_move_id` isn't indexed, defaulting to a Seq.Scan on `account.move`.

## Benchmark
On next.odoo.com, opening an invoice form view

| web_read > read | Before | After  |
|-----------------|--------|--------|
| Timings         | 3.35s  | 0.02ms |

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr