@robodoo 2024-10-25
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/185345.png)](https://mergebot.odoo.com/odoo/odoo/pull/185345)