@pivi-odoo 2024-10-25
Can be backported later to 17.0 (as the field was introduced then)