@pivi-odoo 2024-10-25
@william-andre small PR for a missing index, blocking for internal migration. :)